correction/update dependencies

+++ tests/test-arena.c
@@ -4,13 +4,13 @@
 int main(void) {
     puts("\narena_init");
     arena_t arena = arena_init(0);
-    ASSERT_EQ(0, arena_capacity(arena));
-    ASSERT_EQ(0, arena_total(arena));
+    ASSERT_EQ(0, (int)arena_capacity(arena));
+    ASSERT_EQ(0, (int)arena_total(arena));
 
     puts("\nmalloc_arena initialled");
     void *ptr = arena_alloc(arena, 18);
-    ASSERT_EQ(10240, arena_capacity(arena));
-    ASSERT_EQ(10258, arena_total(arena));
+    ASSERT_EQ(10240, (int)arena_capacity(arena));
+    ASSERT_EQ(10258, (int)arena_total(arena));
 
     char *ptr2 = arena_alloc(arena, 11);
     arena_print(arena);
@@ -25,14 +25,14 @@ int main(void) {
     void *ptr7 = arena_alloc(arena, 2000);
     arena_print(arena);
     void *ptr8 = arena_alloc(arena, 4000);
-    ASSERT_EQ(986, arena_capacity(arena));
-    ASSERT_EQ(10258, arena_total(arena));
+    ASSERT_EQ(986, (int)arena_capacity(arena));
+    ASSERT_EQ(10258, (int)arena_total(arena));
     puts("arena_alloc initialled ended");
 
     puts("\nmalloc_arena(4000) cause total allocation to increase!");
     void *ptr9 = arena_alloc(arena, 4000);
-    ASSERT_EQ(20626, arena_capacity(arena));
-    ASSERT_EQ(24626, arena_total(arena));
+    ASSERT_EQ(20626, (int)arena_capacity(arena));
+    ASSERT_EQ(24626, (int)arena_total(arena));
 
     puts("\nmalloc_arena(4000) again");
     void *ptr10 = arena_alloc(arena, 4000);
@@ -40,7 +40,7 @@ int main(void) {
 
     puts("\n1. arena_clear");
     arena_clear(arena);
-    ASSERT_EQ(20626, arena_capacity(arena));
+    ASSERT_EQ(20626, (int)arena_capacity(arena));
     arena_print(arena);
 
     puts("\nmalloc_aren(6000)a again");
@@ -57,7 +57,7 @@ int main(void) {
 
     puts("\n3. arena_clear");
     arena_clear(arena);
-    ASSERT_EQ(10626, arena_capacity(arena));
+    ASSERT_EQ(10626, (int)arena_capacity(arena));
 
     puts("\nmalloc_arena(1000) x2 again");
     void *ptr13 = arena_alloc(arena, 1000);
@@ -68,8 +68,8 @@ int main(void) {
     puts("\nmalloc_arena(61000)");
     void *ptr15 = arena_alloc(arena, 61000);
     arena_print(arena);
-    ASSERT_EQ(34994, arena_capacity(arena));
-    ASSERT_EQ(95994, arena_total(arena));
+    ASSERT_EQ(34994, (int)arena_capacity(arena));
+    ASSERT_EQ(95994, (int)arena_total(arena));
 
     puts("\nmalloc_arena(7000)");
     void *ptr16 = arena_alloc(arena, 7000);
@@ -77,8 +77,8 @@ int main(void) {
 
     puts("\narena_free");
     arena_free(arena);
-    ASSERT_EQ(0, arena_capacity(arena));
-    ASSERT_EQ(0, arena_total(arena));
+    ASSERT_EQ(0, (int)arena_capacity(arena));
+    ASSERT_EQ(0, (int)arena_total(arena));
     puts("");
 
     return 0;
